reduces dependency with projects_utils (#32)
