+++ services/web/server/tests/unit/isolated/test_projects_utils.py
@@ -11,10 +11,12 @@ import jsonschema
 import pytest
 from jsonschema import ValidationError
 from simcore_service_webserver.projects.project_models import ProjectDict
+from simcore_service_webserver.projects.projects_nodes_utils import (
+    project_get_depending_nodes,
+)
 from simcore_service_webserver.projects.projects_utils import (
     any_node_inputs_changed,
     clone_project_document,
-    project_get_depending_nodes,
 )
